Some minor cleaning up.
