--- src/support/cellmlsupport.cpp
@@ -1,35 +1,0 @@
-/*
-Copyright libOpenCOR contributors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-#include "cellmlsupport.h"
-#include "utils.h"
-
-#include <libcellml>
-
-namespace libOpenCOR::Support {
-
-bool isCellmlFile(const std::string &pFileName)
-{
-    // Try to parse the file contents.
-
-    auto [contents, size] = fileContents(pFileName);
-    auto parser = libcellml::Parser::create();
-    auto model = parser->parseModel(contents.get());
-
-    return parser->issueCount() == 0;
-}
-
-} // namespace libOpenCOR::Support
