--- src/support/combinesupport.cpp
@@ -1,37 +1,0 @@
-/*
-Copyright libOpenCOR contributors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-#include "combinesupport.h"
-
-#include <combine/combinearchive.h>
-
-namespace libOpenCOR::Support {
-
-bool isCombineArchive(const std::string &pFileName)
-{
-    // Try to retrieve a COMBINE archive.
-
-    auto combineArchive = libcombine::CombineArchive();
-    auto res = combineArchive.initializeFromArchive(pFileName);
-
-    if (res) {
-        combineArchive.cleanUp();
-    }
-
-    return res;
-}
-
-} // namespace libOpenCOR::Support
