+++ src/support/support.h
@@ -21,5 +21,7 @@ limitations under the License.
 namespace libOpenCOR::Support {
 
 bool isCellmlFile(const std::string &pFileName);
+bool isCombineArchive(const std::string &pFileName);
+bool isSedmlFile(const std::string &pFileName);
 
 } // namespace libOpenCOR::Support
